requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cssselect==1.1.0
dnspython==2.1.0
fake-useragent==0.1.11
feedparser==6.0.8
filelock==3.3.2
Flask==2.0.2
Flask-PyMongo==2.3.0
Flask-WTF==1.0.0
huggingface-hub==0.1.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.6.4
MarkupSafe==2.0.1
multitasking==0.0.9
numpy==1.21.4
packaging==21.2
pandas==1.3.4
parse==1.19.0
passlib==1.7.4
patsy==0.5.2
Pillow==8.4.0
plotly==5.3.1
plotly-express==0.4.1
psutil==5.8.0
pyee==8.2.2
pymongo==3.12.1
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-html==0.10.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.3
statsmodels==0.13.0
tenacity==8.0.1
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.8.0+cpu
torchaudio==0.8.0
torchvision==0.9.0+cpu
tqdm==4.62.3
transformers==4.12.3
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.2
WTForms==3.0.0
yahoo-fin==0.8.9.1
yfinance==0.1.64
zipp==3.6.0
